test(extension): harden minimap behavior
