+++ packages/extension/src/webview/components/graph/rendering/surface/owned2d/frame/runtime/render.ts
@@ -12,8 +12,9 @@ import type { OwnedGraphLayout } from '../../layout/runtime/model';
 import { advanceOwnedGraphNodeHover, type OwnedGraphNodeHover } from '../../interaction/hover/model';
 import type { OwnedGraphPluginForces } from '../../plugin/forces/model';
 import type { GraphLayoutFixedTimestepClock } from '../../simulation/timing/clock';
-import type { MinimapBounds, MinimapProjection } from '../../minimap/projection';
-import type { MinimapScheduler } from '../../minimap/scheduling';
+import type { MinimapBounds } from '../../minimap/bounds';
+import type { MinimapProjection } from '../../minimap/projection';
+import type { MinimapScheduler } from '../../minimap/state';
 
 export interface OwnedGraphFrameRuntime {
   cameraRef: MutableRefObject<OwnedGraphCamera>; engineStopNotifiedRef: MutableRefObject<boolean>;
