+++ packages/extension/src/webview/components/graph/rendering/surface/owned2d/minimap/scheduling.ts
@@ -1,3 +1,8 @@
+import {
+  completeMinimapRefresh,
+  type MinimapScheduler,
+} from './state';
+
 const MINIMAP_MOVING_REFRESH_INTERVAL_MS = 1000 / 8;
 
 export interface MinimapRefreshInput {
@@ -12,32 +17,6 @@ export interface MinimapRefreshDecision {
   refresh: boolean;
   resetBounds: boolean;
   tightenBounds: boolean;
-}
-
-export interface MinimapScheduler {
-  dirty: boolean;
-  graphIdentity?: object;
-  lastRefreshTimestampMs: number;
-  pendingBoundsReset: boolean;
-  positionVersion: number;
-  styleVersion: number;
-  wasMoving: boolean;
-}
-
-export function createMinimapScheduler(): MinimapScheduler {
-  return {
-    dirty: true,
-    lastRefreshTimestampMs: Number.NEGATIVE_INFINITY,
-    pendingBoundsReset: false,
-    positionVersion: -1,
-    styleVersion: -1,
-    wasMoving: false,
-  };
-}
-
-export function invalidateMinimapScheduler(scheduler: MinimapScheduler): void {
-  scheduler.dirty = true;
-  scheduler.lastRefreshTimestampMs = Number.NEGATIVE_INFINITY;
 }
 
 function observeMinimapChanges(
@@ -64,15 +43,6 @@ function movingCadenceAllowsRefresh(
   return !input.moving
     || input.timestampMs - scheduler.lastRefreshTimestampMs
       >= MINIMAP_MOVING_REFRESH_INTERVAL_MS;
-}
-
-function completeMinimapRefresh(
-  scheduler: MinimapScheduler,
-  timestampMs: number,
-): void {
-  scheduler.dirty = false;
-  scheduler.lastRefreshTimestampMs = timestampMs;
-  scheduler.pendingBoundsReset = false;
 }
 
 export function scheduleMinimapRefresh(
